# Bmatsuo/reduce allocations