--- a/lisp/x/profiler/.gitignore
+++ b/lisp/x/profiler/.gitignore
@@ -1,0 +1,3 @@
+pprof
+pprofout
+callgrind.test_prof