… lisp/eval_test.go
… lisp/lisplib/libjson/libjson_test.go
… lisp/lisplib/libjson/libjson_test.lisp
… lisp/lisplib/libstring/libstring_test.lisp
… lisp/lisplib/libtesting/libtesting.go
… lisp/macro_test.go
… lisp/op_test.go