The diff looks a bit all over the place. But, looking at the individual commits would give a more focused view of the changes.